requirements.txt
connexion[flask,uvicorn,swagger-ui]
flask-cors
PyYAML
kafka-python
sqlalchemy
mysqlclient
httpx
apscheduler
cryptography